+++ .github/CHANGELOG.md
@@ -17,6 +17,31 @@
   | `'tf'`       | `tuple(arrays)`    | `array`               | `tuple(array)`            |
   | `'torch'`    | `tuple(arrays)`    | `array`               | `tuple(array)`            |
 
+  .. list-table:: Output format of ``classical_jacobian``
+     :widths: 25 25 25 25
+     :header-rows: 1
+
+     * - Interface
+       - ``argnums=None``
+       - ``type(argnums)=int``
+       - ``argnums=Sequence[int]``
+     * - ``'autograd'``
+       - ``tuple(arrays)``\*
+       - ``array``
+       - ``tuple(array)``
+     * - ``'jax'``
+       - ``array``
+       - ``array``
+       - ``tuple(array)``
+     * - ``'tf'``
+       - ``tuple(arrays)``
+       - ``array``
+       - ``tuple(array)``
+     * - ``'torch'``
+       - ``tuple(arrays)``
+       - ``array``
+       - ``tuple(array)``
+
   \* If all QNode argument are scalars, the tuple is unpacked and the one-dimensional Jacobian
   arrays are stacked into one `array`. This is due to the behaviour of `qml.jacobian`.